Backend Curation API 1차 완성

See merge request s11-webmobile2-sub2/S11P12A506!13

+++ backend/curation-service/src/main/java/com/ssafy/bookkoo/curationservice/repository/CurationRepository.java
@@ -3,6 +3,6 @@ package com.ssafy.bookkoo.curationservice.repository;
 import com.ssafy.bookkoo.curationservice.entity.Curation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CurationRepository extends JpaRepository<Curation, Integer> {
+public interface CurationRepository extends JpaRepository<Curation, Long> {
 
 }
